+++ plugin/includes/admin/class-merchants-screen.php
@@ -178,6 +178,10 @@ class Supcomp_Merchants_Screen {
 									<?php wp_nonce_field( self::NONCE_STAT . '_' . $r->id ); ?>
 									<?php submit_button( $next_label, 'link-delete', 'submit', false ); ?>
 								</form>
+								<?php if ( Supcomp_Deletion_Service::merchant_is_deletable( $r ) ) : ?>
+									&nbsp;|&nbsp;
+									<a href="<?php echo esc_url( Supcomp_Deletion_Admin::confirm_url( 'merchant', (int) $r->id ) ); ?>" style="color:#a00"><?php esc_html_e( 'Delete', 'supplement-compare' ); ?></a>
+								<?php endif; ?>
 							</td>
 						</tr>
 					<?php endforeach; ?>